big ol refactor
- separated files
- es6 classes
- template strings
- weback -> rollup
- iife + cjs build
- babelrc, eslintrc, npmignore
