+++ .babelrc
@@ -1,0 +1,18 @@
+{
+	"env": {
+		"test": {
+			"presets": ["env"]
+		}
+	},
+	"presets": [
+		[
+			"env",
+			{
+				"modules": false
+			}
+		]
+	],
+	"plugins": [
+		"transform-class-properties"
+	]
+}
