+++ .eslintrc
@@ -1,0 +1,20 @@
+{
+	"extends": "eslint:recommended",
+	"env": {
+		"browser": true,
+		"jest": true,
+		"es6": true
+	},
+	"parserOptions": {
+		"ecmaVersion": 6,
+		"sourceType": "module",
+		"ecmaFeatures": {
+			"modules": true,
+			"experimentalObjectRestSpread": true
+		}
+	},
+	"rules": {
+		"no-console": 1,
+		"no-unused-vars": 1
+	}
+}
