+++ src/main.js
@@ -1,0 +1,2 @@
+import WebGLazy from './WebGLazy';
+export default WebGLazy;
